moved DoWithTool to Runtime
