+++ ModernRonin.ProjectRenamer/Executor.cs
@@ -1,7 +1,4 @@
 ﻿using System;
-using System.ComponentModel;
-using System.Diagnostics;
-using System.IO;
 
 namespace ModernRonin.ProjectRenamer
 {
@@ -17,14 +14,6 @@ namespace ModernRonin.ProjectRenamer
         public string DotNetRead(string arguments) =>
             ToolRead(ToolDotnet, arguments, () => StandardErrorHandler(ToolDotnet, arguments));
 
-        public void Git(string arguments) => Tool(ToolGit, arguments);
-
-        public void Git(string arguments, Action onNonZeroExitCode) =>
-            Tool(ToolGit, arguments, onNonZeroExitCode);
-
-        public string GitRead(string arguments) =>
-            ToolRead(ToolGit, arguments, () => StandardErrorHandler(ToolGit, arguments));
-
         public void Error(string msg, bool doResetGit = false)
         {
             Console.Error.WriteLine(msg);
@@ -37,44 +26,15 @@ namespace ModernRonin.ProjectRenamer
             Runtime.Abort();
         }
 
+        public void Git(string arguments) => Tool(ToolGit, arguments);
+
+        public void Git(string arguments, Action onNonZeroExitCode) =>
+            Tool(ToolGit, arguments, onNonZeroExitCode);
+
+        public string GitRead(string arguments) =>
+            ToolRead(ToolGit, arguments, () => StandardErrorHandler(ToolGit, arguments));
+
         public void RollbackGit() => Git("reset --hard HEAD", () => { });
-
-        static void DoWithTool(string tool,
-            string arguments,
-            Action onNonZeroExitCode,
-            Action<ProcessStartInfo> configure,
-            Action<Process> onSuccess)
-        {
-            var psi = new ProcessStartInfo
-            {
-                FileName = tool,
-                Arguments = arguments,
-                UseShellExecute = false,
-                CreateNoWindow = false
-            };
-            configure(psi);
-            try
-            {
-                var process = Process.Start(psi);
-                process.WaitForExit();
-                if (process.ExitCode != 0) onNonZeroExitCode();
-                else onSuccess(process);
-            }
-            catch (Win32Exception)
-            {
-                onProcessStartProblem();
-            }
-            catch (FileNotFoundException)
-            {
-                onProcessStartProblem();
-            }
-
-            void onProcessStartProblem()
-            {
-                Console.Error.WriteLine($"{tool} could not be found - make sure it's on your PATH.");
-                onNonZeroExitCode();
-            }
-        }
 
         void StandardErrorHandler(string tool, string arguments)
         {
@@ -83,7 +43,7 @@ namespace ModernRonin.ProjectRenamer
 
         static void Tool(string tool, string arguments, Action onNonZeroExitCode)
         {
-            DoWithTool(tool, arguments, onNonZeroExitCode, psi => psi.RedirectStandardOutput = false,
+            Runtime.DoWithTool(tool, arguments, onNonZeroExitCode, psi => psi.RedirectStandardOutput = false,
                 _ => { });
         }
 
@@ -93,7 +53,7 @@ namespace ModernRonin.ProjectRenamer
         static string ToolRead(string tool, string arguments, Action onNonZeroExitCode)
         {
             var result = string.Empty;
-            DoWithTool(tool, arguments, onNonZeroExitCode, psi => psi.RedirectStandardOutput = true,
+            Runtime.DoWithTool(tool, arguments, onNonZeroExitCode, psi => psi.RedirectStandardOutput = true,
                 p => result = p.StandardOutput.ReadToEnd());
             return result;
         }
